Fix Hardy revert for pre-3.3.0 Characters

+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/character/effects/EffectFactory.kt
@@ -16,7 +16,7 @@ class EffectFactory {
                 val name = item.talent.name.trim()
 
                 listOfNotNull(
-                    HardyWoundsModification.fromTalentOrNull(name, item.talent.taken.toUInt()),
+                    HardyWoundsModification.fromTalentOrNull(name, item.talent.taken),
                     CharacteristicChange.fromTalentNameOrNull(name),
                 )
             }
